requirements.txt
flask==2.2.3
flask-cors==3.0.10
pymysql==1.0.3
python-dotenv==1.0.0
gunicorn==20.1.0
cryptography==40.0.1